% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solarmanpv.R
\name{plantlist}
\alias{plantlist}
\title{List available "plants" (installations)}
\usage{
plantlist(token, ...)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
List available "plants" (installations)
}
