% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solarmanpv.R
\name{station_call}
\alias{station_call}
\title{Generic API call once you have token}
\usage{
station_call(token, route, ...)
}
\arguments{
\item{...}{}
}
\value{
endpoint return
}
\description{
Generic API call once you have token
}
